+++ Change_log.md
@@ -1,5 +1,6 @@
 # [main](https://github.com/jzhang-github/AGAT/tree/main)
 - Fix warning when using `torch.load`. [agat/lib/model_lib.py](https://github.com/jzhang-github/AGAT/blob/main/agat/lib/model_lib.py#L42-L44)
+- Fix bugs is `ase.atoms` has `ase.constraints.FixScaled` and `ase.constraints.FixedLine`.
 
 # [v9.0.0](https://github.com/jzhang-github/AGAT/tree/v9.0.0)
 **Note: AGAT after this version (included) cannot load the well-trained model before.** If you need to do so, please use v8.0.5: https://pypi.org/project/agat/8.0.5/